+++ scripts/spack/packages/tribol/package.py
@@ -1,7 +1,7 @@
-# Copyright 2013-2023 Lawrence Livermore National Security, LLC and other
-# Spack Project Developers. See the top-level LICENSE file for details.
+# Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
+# other Tribol Project Developers. See the top-level COPYRIGHT file for details.
 #
-# SPDX-License-Identifier: (Apache-2.0 OR MIT)
+# SPDX-License-Identifier: (MIT)
 
 import os
 import socket
@@ -57,6 +57,8 @@ class Tribol(CachedCMakePackage, CudaPackage, ROCmPackage):
             description="Build with portable kernel execution support")
     variant("openmp",   default=False,
             description="Build with OpenMP support")
+    variant("enzyme",   default=False,
+            description="Build with Enzyme support")
 
     # -----------------------------------------------------------------------
     # Dependencies
@@ -74,6 +76,8 @@ class Tribol(CachedCMakePackage, CudaPackage, ROCmPackage):
 
     depends_on("raja@2024.02.0:", when="+raja")
     depends_on("umpire@2024.02.0:", when="+umpire")
+
+    depends_on("enzyme", when="+enzyme")
     
     depends_on("axom+raja", when="+raja")
     depends_on("axom~raja", when="~raja")
@@ -146,6 +150,8 @@ class Tribol(CachedCMakePackage, CudaPackage, ROCmPackage):
 
     conflicts("+cuda", when="+rocm")
     conflicts("+openmp", when="+rocm")
+
+    requires("%clang", when="+enzyme")
 
     def _get_sys_type(self, spec):
         sys_type = spec.architecture
@@ -385,7 +391,7 @@ class Tribol(CachedCMakePackage, CudaPackage, ROCmPackage):
                                             dep_dir))
 
         # optional tpls
-        for dep in ('raja', 'umpire'):
+        for dep in ('raja', 'umpire', 'enzyme'):
             if spec.satisfies('^{0}'.format(dep)):
                 dep_dir = get_spec_path(spec, dep, path_replacements)
                 entries.append(cmake_cache_path('%s_DIR' % dep.upper(),