+++ src/redecomp/transfer/MatrixTransfer.cpp
@@ -304,7 +304,6 @@ MPIArray<int> MatrixTransfer::buildRecvTrialElemDofs( const RedecompMesh& trial_
 {
   auto recv_trial_elem_dofs = MPIArray<int>( &getMPIUtility() );
 
-  auto rank = getMPIUtility().MyRank();
   auto n_ranks = getMPIUtility().NRanks();
 
   // List of trial element offsets sorted by the parent test space rank and the
@@ -385,8 +384,7 @@ MPIArray<int> MatrixTransfer::buildRecvTrialElemDofs( const RedecompMesh& trial_
           recv_trial_elem_offsets[src_trial_r] = std::move( send_vals );
         } );
     // send parent trial vdofs back to test redecomp rank
-    auto dof_offsets = parent_trial_fes_.GetDofOffsets();
-    auto first_dof = HYPRE_AssumedPartitionCheck() ? dof_offsets[0] : dof_offsets[rank];
+    auto first_dof = parent_trial_fes_.GetMyDofOffset();
     auto trial_dofs_by_rank = MPIArray<int>( &getMPIUtility() );
     trial_dofs_by_rank.SendRecvEach(
         [this, &trial_p2r_elems, &recv_trial_elem_offsets, first_dof]( axom::IndexType src_trial_r ) {