+++ src/tests/tribol_common_plane_gpu.cpp
@@ -98,7 +98,7 @@ double runExample( int num_elems_1d )
       break;
 #endif
 #ifdef TRIBOL_USE_HIP
-    case MemorySpace::Hip:
+    case MemorySpace::Device:
       device.Configure( "hip" );
       break;
 #endif